+++ projects/graphics/CentralRenderer/sources/IndexBuffer.h
@@ -5,15 +5,15 @@
 class IndexBuffer
 {
 public:
-	IndexBuffer(const uint16_t* data, size_t count);
+	IndexBuffer(const uint32_t* data, uint32_t count);
 	~IndexBuffer();
 
 	void bind() const;
 	void unbind() const;
 
-	inline size_t getCount() const { return m_count; };
+	inline uint32_t getCount() const { return m_count; };
 
 private:
 	uint32_t m_rendererID;
-	size_t m_count;
+	uint32_t m_count;
 };
